+++ db/migrate/20180212124737_create_suppliers.rb
@@ -2,6 +2,7 @@ class CreateSuppliers < ActiveRecord::Migration[5.1]
   def change
     create_table :suppliers do |t|
       t.string :name
+      t.string :email
       t.text :address
       t.string :contact_person
       t.string :contact_number